Move FeatureFlagOverrideStore to its own file
